(split)Bug: none
Selenium Harness:  for the auth token timeout test, increase the timeout for the coverage tests.  The dev=1&debug=0 flags make the login much slower.  So, the increase of timeout is needed so that the client won't send a SOAP request after the timeout has occurred (thus putting the user into the login screen before "send" is pressed).

1. Login
2. Get main screen
3. Compose message
4. A SaveDraftRequest is fired
5. Auth token expires
6. Press send
7. User is placed in login

For coverage, the delay between step 1 and step 2 is over 20+ seconds.  Therefore, step 5 happens before step 4 ... sending the user to step 7 before the harness executes step 6.

Copied from Perforce
 Change: 401352