(split)Bug: none
Selenium Harness: compose in new window tests sometimes fail because the message isn't delivered to the postfix queue before the harness checks it.  Since the new window is closed on clicking "Send", there is no 'busy overlay' to allow sync between the client and harness.  Add a SearchRequest check to make sure the message is delivered to the end point.  Loop for 30 seconds for delivery.

Copied from Perforce
 Change: 401905