(split)Bug: 63353
Selenium Harness: update dialog ID for "The selected time is in the past. Message will be sent immediately."

In main, the dialog id is <div id='OkCancel' .../>
In 8.x, the dialog id is <div id='ShowDelayPastDialog' .../>

Either way, there is a unique ID, so no app changes are required.  Updating the mainline branch to expect OkCancel.

http://bugzilla.zimbra.com/show_bug.cgi?id=63353

Copied from Perforce
 Change: 415863