(split)Bug: none
Selenium Harness: change basic external account registration from sanity to smoke.  it is debatable whether the test case matches the sanity definition (create basic objects, i.e. external account).  and the test does not work in the dev environment, which gives a failure in the sanity results that dev is attempting to run

Copied from Perforce
 Change: 439078