(split)Bug: none
Selenium Harness: for any tests that create account in the constructor, move the account creation to the test cases.  If the create account fails in the constructor (even for a test that doesn't require the account), the entire suite will not be executed (which is bad).

Found another instance.  Follow up to Change 453757 by rhoades@RHOADES-7-SELNG on 2013/03/14 11:04:45

Copied from Perforce
 Change: 454019