(split)Bug: none
Selenium Harness: move zCheckboxSet from PagePreferences to AbsSeleniumObject.

It is good to have a generic method that sets a checkbox status to a particular state (true=checked, false=unchecked) no matter what the starting state of the checkbox is.  This is different than sClick, which toggles the value, which may be the wrong action if the test case wants a particular end result.

Copied from Perforce
 Change: 461933