Fixed some of the probable issues for skip and some improvements.
Added small sleep in zWaitForWorkInProgressDialogInVisible() to make
sure that it captures the dialog reliably.

PageMain.java: Kept only the refresh code in zHandleDialog(); This is
required for every test as beforeMethod has a starting page whose
zNavigateTo() method only handles navigation from the home page. Because
of this the page needs to be refreshed after/before every test to bring
back the application home page.

AdminCommonTest.java: Commented out the part of code as refreshing the
page will always bring back the application to main page.